py_fanza_auto/requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
APScheduler>=3.10.0
selenium>=4.15.0
webdriver-manager>=4.0.0
